+++ x/gov/src/types/vote.rs
@@ -15,7 +15,7 @@ impl Vote {
         [
             KEY_VOTES_PREFIX.as_slice(),
             &self.proposal_id.to_be_bytes(),
-            &[self.voter.len() as u8], // We save 'cause `AccAddress` len shoudn't be bigger than 255
+            &[self.voter.len()],
             self.voter.as_ref(),
         ]
         .concat()
